Decoupled Builder and Fetcher
